Adapt the related keyphrase upsell under SEO analysis to be the same
